feat: refactor polygon curve
